+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Add `partitioned` option
+
 0.5.0 / 2022-04-11
 ==================